Moves scaffolding under pkg folder
