+++ cli/pkg/scaffolding/git_test.go
