+++ cli/pkg/scaffolding/scaffold.go
